@@ -15,31 +15,40 @@
   },
   "main": "node_modules/expo/AppEntry.js",
   "dependencies": {
-    "@expo/vector-icons": "^10.0.0",
+    "@expo/vector-icons": "^12.0.0",
+    "@react-native-async-storage/async-storage": "^1.13.0",
+    "@react-native-community/masked-view": "0.1.10",
+    "@react-native-community/netinfo": "6.0.0",
+    "@react-navigation/compat": "^5.3.15",
+    "@react-navigation/drawer": "^5.12.5",
+    "@react-navigation/native": "^5.9.4",
+    "@react-navigation/stack": "^5.14.5",
     "axios": "^0.21.1",
-    "expo": "^35.0.0",
-    "expo-asset": "~7.0.0",
-    "expo-constants": "~7.0.0",
-    "expo-font": "~7.0.0",
-    "jest-expo": "^35.0.0",
+    "expo": "^41.0.0",
+    "expo-app-loading": "^1.0.3",
+    "expo-asset": "~8.3.1",
+    "expo-constants": "~10.1.3",
+    "expo-font": "~9.1.0",
+    "expo-splash-screen": "~0.10.2",
     "moment": "^2.22.2",
     "prop-types": "^15.6.2",
-    "react": "16.8.3",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
+    "react": "16.13.1",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-41.0.0.tar.gz",
     "react-native-collapsible": "^1.0.0",
-    "react-native-gesture-handler": "~1.3.0",
-    "react-native-maps": "~0.25.0",
-    "react-native-popup-dialog": "^0.15.0",
-    "react-native-reanimated": "~1.2.0",
+    "react-native-gesture-handler": "~1.10.2",
+    "react-native-maps": "0.27.1",
+    "react-native-popup-dialog": "0.15.0",
+    "react-native-reanimated": "~2.1.0",
     "react-native-root-toast": "^3.0.1",
-    "react-native-simple-store": "^1.3.0",
-    "react-native-svg": "~9.9.2",
-    "react-native-textinput-effects": "^0.4.0",
-    "react-navigation": "^3.12.0",
+    "react-native-safe-area-context": "3.2.0",
+    "react-native-screens": "~3.0.0",
+    "react-native-svg": "12.1.0",
+    "react-native-textinput-effects": "^0.6.1",
+    "react-native-webview": "11.2.3",
     "react-redux": "^5.0.7",
     "redux": "^3.7.2",
     "redux-persist": "^5.10.0",
-    "sentry-expo": "~2.0.0"
+    "@react-native-picker/picker": "1.9.11"
   },
   "devDependencies": {
     "prettier": "^1.14.2"
