@@ -1,5 +1,6 @@
 import React from 'react';
-import { ActivityIndicator, AsyncStorage, NetInfo, SectionList, Text, TouchableOpacity, View } from 'react-native';
+import { ActivityIndicator, SectionList, Text, TouchableOpacity, View } from 'react-native';
+import AsyncStorage from '@react-native-async-storage/async-storage'
 import axios from 'axios';
 import { Hideo } from 'react-native-textinput-effects';
 import { connect } from 'react-redux';
@@ -7,50 +8,16 @@ import moment from 'moment';
 import { FontAwesome, MaterialCommunityIcons } from '@expo/vector-icons';
 import Toast from 'react-native-root-toast';
 
-import NavBarHelper from '../components/NavBarHelper';
 import SectionListHeader from '../components/ui/SectionListHeader';
 import Split from '../components/ui/Split';
 import GroupRow from '../components/GroupRow';
 import style from '../Style';
 import Translator from '../utils/translator';
+import DeviceUtils, { AppContext } from '../utils/DeviceUtils';
 
 class Home extends React.Component {
-    static navigationOptions = ({ navigation, navigationOptions, screenProps }) => {
-        const title = Translator.get('GROUPS');
-        const leftButton = (
-            <TouchableOpacity
-                onPress={() => {
-                    navigation.openDrawer();
-                }}
-                style={{
-                    justifyContent: 'space-around',
-                    paddingLeft: 16,
-                }}>
-                <View
-                    style={{
-                        flexDirection: 'row',
-                        justifyContent: 'space-between',
-                    }}>
-                    <MaterialCommunityIcons
-                        name="menu"
-                        size={32}
-                        style={{
-                            color: '#F0F0F0',
-                            height: 32,
-                            width: 32,
-                        }}
-                    />
-                </View>
-            </TouchableOpacity>
-        );
-
-        return NavBarHelper({
-            title,
-            headerLeft: leftButton,
-            themeName: screenProps.themeName,
-        });
-    };
-
+    static contextType = AppContext;
+    
     constructor(props) {
         super(props);
         this.state = {
@@ -91,7 +58,7 @@ class Home extends React.Component {
                     key: previousSection,
                     data: [],
                     sectionIndex: ++sectionIndex,
-                    colorIndex: sectionIndex % style.Theme[this.props.themeName].sections.length,
+                    colorIndex: sectionIndex % style.Theme[this.context.themeName].sections.length,
                 };
             }
 
@@ -132,8 +99,7 @@ class Home extends React.Component {
     async fetchList() {
         let list = null;
 
-        const isConnected = (await NetInfo.getConnectionInfo()) !== 'none';
-        if (isConnected) {
+        if (await DeviceUtils.isConnected()) {
             try {
                 const response = await axios.get('https://hackjack.info/et/json.php?clean=true');
                 this.setState({ cacheDate: null });
@@ -209,7 +175,7 @@ class Home extends React.Component {
     }
 
     render() {
-        const theme = style.Theme[this.props.themeName];
+        const theme = style.Theme[this.context.themeName];
 
         let content = null,
             cache = null;
